+++ config/test.exs
@@ -4,6 +4,7 @@ config :algoliax,
   application_id: "APPLICATION_ID",
   api_key: "api_key",
   batch_size: 1,
+  http_client: Algoliax.Support.HttpClient,
   ecto_repos: [Algoliax.Repo]
 
 config :algoliax, Algoliax.Repo,